pyproject.toml
[project]
name = "plastex_lpil_plugin"
version = "0.0.1"
description = "A PlasTeX plugin describing the LPiL package"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "plastex>=3.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[project.entry-points."plastex.plugin"]
lpilPasTeXPlugin = "lpilPlasTeXPlugin:10"
